+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 Improvements:
 * GET `/tradeItems` now supports a repeatable `id` query parameter to filter trade items by their identifiers.
+* [OLMIS-8118](https://openlmis.atlassian.net/browse/OLMIS-8118): Reject kit child quantity exceeding Integer max.
 
 15.5.0 / 2026-06-09
 ==================